pyproject.toml
[project]
name = "telegram-claude-bot"
version = "1.0.0"
description = "Telegram bot interface for Claude Code CLI"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "cron-descriptor>=2.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
telegram-claude-bot = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
cache_dir = ".build/.pytest_cache"
markers = [
    "integration: 통합 테스트 (텔레그램 목킹, 실제 Repository 사용)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
